oneplus3: Hex-edit camera HAL to fix FFC crashes on the OP3T Capturing images with the OP3T's front-facing camera at resolutions above 8 MP results in the camera subsequently crashing.
The camera HAL loads different settings based on the name of the app that's
using the camera, but since the code that sends app names to the camera HAL
is in the OxygenOS framework, this system is broken on custom ROMs and
yields the dysfunctional front-facing camera.

In order to fix this, hex-edit the camera HAL to make it think that
"com.oneplus.camera" is always in use. The changes are as follows:
At file offset 0xcbf52, the four bytes "78 F7 12 EE" are replaced with
"C0 46 00 20". The "78 F7 12 EE" corresponds to a call to strncmp(), and
the return value of the call is stored in register r0. After the strncmp()
call, there is a comparison between r0 and 0 and a conditional jump after
that. In order to emulate "com.oneplus.camera", the "C0 46 00 20" replacing
the strncmp() call consists of a nop instruction (C0 46) and a movs
instruction (00 20) to move a literal 0 into register r0.

A huge thanks to Thecrazyskull for providing crucial info that led to this
solution.